feat(2026-03-baseball): refactor shell to reuse in other components
